file transfer handlers
This commit implements actual file movements. It implements simple
file IO using stdio methods and transfers files in 8k blocks.

Notice in the receiving of the file that it is possible (and likely
if not hitting localhost) that there is file part being sent with the
handshake message, that is handled with that code with the rather
extreme pointer arithmetic in the line 275.

Signed-off-by: Zeljko <zeljko@zwr.fi>