print binary received file to stdout
Previous approach with printf would print until the first zero
byte, which was OK while testing with text file but was of course
making rubbish on binary files. Using fwrite instead quickly solves
the issue.

This commit passes the test with large files, exchanging large photos
of Pamela Anderson and Claudia Schiffer, which should be sufficient
for the purpose of this work.

Signed-off-by: Zeljko <zeljko@zwr.fi>